+++ medusa/app.py
@@ -323,6 +323,9 @@ class MedusaApp(object):
         self.SKIP_REMOVED_FILES = False
         self.ALLOWED_EXTENSIONS = ['srt', 'nfo', 'sub', 'idx']
 
+        self.FFMPEG_CHECK_STREAMS = False
+        self.FFMPEG_PATH = ''
+
         self.NZBS = False
         self.NZBS_UID = None
         self.NZBS_HASH = None