+++ themes-default/slim/src/store/modules/config/postprocessing.js
@@ -42,6 +42,10 @@ const state = {
         minFrequency: null,
         torrentSeedRatio: null,
         torrentSeedAction: null
+    },
+    ffmpeg: {
+        checkStreams: null,
+        path: ''
     }
 };